Almost ready
It's basically ready, there are only two things left: create the energy saving mode profile and test it further. The first test I'm going to do is try to make the battery last 10 hours with my usual use and improve it to 11 hours. If I reach that time, I'll post it immediately.